\name{cd}
\alias{cd}
\title{
Change directory.
}
\description{
Change directory to the one corresponding to the row number listed by ``f()''.
}
\usage{
cd(dir)
}
\arguments{
  \item{dir}{
The row number corresponding to the directory listed by ``f()''.
}
}
\value{
Change to new directory.
}
\author{
Antoine Tremblay, IWK Health Center, <trea26@gmail.com>
}
\seealso{
\code{\link{f}}
\code{\link{cdf}}
\code{\link{cdup}}
\code{\link{setwd}}
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
