\name{cdf}
\alias{cdf}
\title{Change directory; list files and directories in new directory using function \code{f}.}
\description{Change directory to the one corresponding to the row number returned by function \code{f}.}
\usage{cdf(dir)}
\arguments{\item{dir}{The row number corresponding to the directory listed by function \code{f}.}
}
\value{Cheange to new directory and list files and directories in new directory using function \code{f}.}
\author{Antoine Tremblay, IWK Health Center, \email{trea26@gmail.com}}
\seealso{
\code{\link{f}}
\code{\link{cd}}
\code{\link{cdup}}
\code{\link{setwd}}
}
