\name{cdup}
\alias{cdup}
\title{
Change directory one level up. 
}
\description{
Change directory one level up and list directory and files in new directory.
}
\usage{
cdup()
}
\value{
Change directory one level up.
}
\author{
Antoine Tremblay, IWK Health Center, <trea26@gmail.com>
}

\seealso{
\code{\link{f}; }
\code{\link{cd}; }
\code{\link{cdf}; }
\code{\link{setwd} }
}
