\name{cd}
\alias{cd}
\title{Change directory.}
\description{Change directory to the one corresponding to the row number listed by function \code{f}.}
\usage{cd(dir)}
\arguments{\item{dir}{The row number corresponding to the directory list returned by function \code{f}.}}
\value{Change directory to the selected one.}
\author{Antoine Tremblay, IWK Health Center, \email{trea26@gmail.com}}
\seealso{
\code{\link{f}; }
\code{\link{cdf}; }
\code{\link{cdup}; }
\code{\link{setwd} }
}
