\name{cdf}
\alias{cdf}
\title{
Change directory and list files and directories in new directory using function ``f()''.
}
\description{
Change directory to the one corresponding to the row number listed by ``f()'' and directories in new directory using function ``f()''.
}
\usage{
cdf(dir)
}
\arguments{
  \item{dir}{
The row number corresponding to the directory listed by ``f()''.
}
}
\value{
Change to new directory and list files and directories in new directory using ``f()''.
}
\author{
Antoine Tremblay, IWK Health Center, <trea26@gmail.com>
}

\seealso{
\code{\link{f}}
\code{\link{cd}}
\code{\link{cdup}}
\code{\link{setwd}}
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
