\name{cn}
\alias{cn}
\title{List the column names of a data frame in matrix format.}
\description{The colum names of the specified data frame are listed in matrix format, that is, each one appears in one row preceded by the row number.}
\usage{cn(data.frame)}
\arguments{\item{data.frame}{A data frame.}}
\value{A matrix containing the column names of the data frame.}
\author{Antoine Tremblay, Statistics Canada, \email{trea26@gmail.com}}
\seealso{
\code{\link{colnames}}
}
